requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
tenacity>=8.2.3
PyJWT>=2.8.0
urllib3>=2.1.0
supabase>=2.3.0
openai>=1.6.0
